add Service code

+++ Springhibernate/src/main/java/quyet/demo/spring/model/Store.java
@@ -17,7 +17,7 @@ public class Store {
     this.addresses = addresses;
   }
 
-  public long getStoreId() {
+  public int getStoreId() {
     return storeId;
   }
